+++ Kernel/file_system/fat32_calc.h
@@ -7,5 +7,5 @@ uint64_t fat32_calc_get_fat_size_in_sectors();
 uint64_t fat32_calc_get_fat_start_lba();
 uint32_t fat32_calc_get_fat_sectors_per_cluster();
 void fat32_calc_init_parameters(fat32_bpb_t* bpb, uint64_t partition_start_lba);
-void fat32_calc_fat_table_offsets(uint32_t results[], uint32_t cluster_no);
-void fat32_calc_dir_offsets(uint32_t results[], uint32_t cluster_no);
+void fat32_calc_fat_table_offsets(uint64_t results[], uint32_t cluster_no);
+void fat32_calc_dir_offsets(uint64_t results[], uint32_t cluster_no);
